requirements.txt
pyfiglet==0.8.post1
pillow==8.0.1